backend/requirements.txt
# FastAPI & Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Authentication - LDAP / Active Directory
ldap3==2.9.1
python-jose[cryptography]==3.3.0

# Security
bcrypt>=4.0.0
itsdangerous==2.2.0

# HTTP Client (for Ollama)
httpx==0.27.0

# JSON (fast serialisation for streaming)
orjson>=3.9.0

# Document text extraction
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dotenv==1.0.1
orjson==3.10.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
